Creates a main.rs file
The `path = "src/main.rs"` in the `Cargo.toml` file is where the code for your first nickel app resides. You may want to choose a different file name but at the time of writing Cargo assumes your file to be in `src/main.rs` for the `cargo run` command to work.